Inline adapter policy helpers into WDDM device header
