+++ CMakeLists.txt
@@ -119,7 +119,6 @@ set ( ROCDXG_SRC "src/debug.cpp"
                  "src/hsakmtmodel.cpp"
                  "src/dxcore_loader.cpp"
                  "src/ais.cpp"
-                 "src/wddm/adapter_policy.cpp"
                  "src/wddm/device.cpp"
                  "src/wddm/gpu_memory.cpp"
                  "src/wddm/va_mgr.cpp"
