+++ tests/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_executable(adapter_policy_test
-  adapter_policy_test.cpp
-  ${PROJECT_SOURCE_DIR}/src/wddm/adapter_policy.cpp)
+  adapter_policy_test.cpp)
 
 target_include_directories(adapter_policy_test PRIVATE
   ${PROJECT_SOURCE_DIR}/include
