+++ tests/adapter_policy_test.cpp
@@ -2,7 +2,7 @@
 #include <cstdlib>
 #include <cstring>
 
-#include "wddm/adapter_policy.h"
+#include "impl/wddm/device.h"
 
 namespace {
